+++ rest/src/main/java/io/github/gprindevelopment/rest/common/ComponenteClient.java
@@ -61,7 +61,7 @@ public class ComponenteClient {
         return Integer.parseInt(valorCabecalho);
     }
 
-    protected <T> RespostaCamara<T> consultarPorId(int id, String urlBase, Type tipoEsperado) throws RespostaNaoEsperadaException, CamaraClientStatusException, IOException {
+    protected <T> RespostaCamara<T> consultarPorId(long id, String urlBase, Type tipoEsperado) throws RespostaNaoEsperadaException, CamaraClientStatusException, IOException {
         Request requisicao = new RequisicaoBuilder(urlBase).segmentosPath(String.valueOf(id)).build();
         Call chamada = client.newCall(requisicao);
         return executarChamada(chamada, tipoEsperado);
